refactor: move extract_path_segments from analytics to lib
It's pure path parsing with no analytics coupling, and lib is its only caller
(feeding both the analytics event and the location broadcast). Move it next to
the sibling request helpers (header_str/extract_request_id); analytics.rs is
now purely the Analytics Engine sink.
